requirements.txt
# PromptLab Backend Dependencies
# Core Flask framework and extensions
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.36

# Utilities
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0

# Production server
gunicorn==21.2.0

# Environment management
python-dotenv==1.0.0